backend/pyproject.toml
[project]
name = "vectorsight"
version = "0.1.0"
description = "SVG spatial analysis engine — geometry transforms for LLM comprehension"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "svgpathtools>=1.6.1",
    "lxml>=5.3.0",
    "numpy>=2.2.0",
    "scipy>=1.15.0",
    "scikit-learn>=1.6.0",
    "scikit-image>=0.25.0",
    "networkx>=3.4.0",
    "shapely>=2.0.7",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "app"
module-root = ""

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
